Further establishing the "atom" type.
For the time being, "atom" will be the unit for the length
of vectors. If we're serious about separating vectors of
atoms from vectors of references to save memory, the
allocator will have to be changed to deliver each type with
the correct alignment: Until then, we might as well say
"typedef intptr_t atom".

+++ gc.h
@@ -25,6 +25,7 @@
 #define EDEN_OVERHEAD 5 // For the sake of testing.
 
 #include <gmp.h> // For the definition of "mpz_t".
+#include <stdint.h> // For the definition of "intptr_t".
 
 typedef struct vectorStruct {
   struct vectorStruct *prev, *next;
@@ -33,7 +34,7 @@ typedef struct vectorStruct {
 } *vector;
 
 typedef vector obj;
-typedef int atom;
+typedef intptr_t atom;
 #define CELLS_REQUIRED_FOR_BYTES(n) (((n) + sizeof(atom) - 1) / sizeof(atom))
 
 vector shelteredValue(vector);
